Creating components home and toolbar
